puppet banner

+++ drupal/manifests/init.pp
@@ -21,7 +21,7 @@ class drupal {
     }
 
     file { "/tmp/db_settings.inc" :
-        ensure => present,
+        ensure  => present,
         content => template("drupal/db_settings.inc.erb"),
         require => File["/vagrant/htdocs/sites/default/settings.php"],
     }
@@ -31,13 +31,20 @@ class drupal {
         require => File["/tmp/db_settings.inc"],
     }
 
+    exec { "add_puppet_banner" :
+        command => 'sed -i "1a\\\n##### Managed by Puppet - Intracto Vagrant #####################################" /vagrant/htdocs/sites/default/settings.php',
+        require => File['/vagrant/htdocs/sites/default/settings.php'],
+    }
+
     exec { "get-drush" :
         command => 'curl http://ftp.drupal.org/files/projects/drush-7.x-5.8.tar.gz | tar xzvf - -C /opt',
         require => File["/tmp/db_settings.inc"],
+        unless  => 'test -f /usr/local/bin/drush',
     }
 
     exec { "activate-drush" :
         command => 'ln -s /opt/drush/drush /usr/local/bin ; chmod -R 777 /opt/drush/lib',
         require => Exec["get-drush"],
+        unless  => 'test -f /usr/local/bin/drush',
     }
 }
